+++ lib/audioplayers.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 import 'dart:io';
+import 'dart:typed_data';
 import 'dart:ui';
 
 import 'package:flutter/foundation.dart';
@@ -67,6 +68,11 @@ enum PlayerMode {
   /// Also, it is not possible to use the seek method to set the audio a
   /// specific position.
   LOW_LATENCY
+}
+
+enum PlayerControlCommand {
+  NEXT_TRACK,
+  PREVIOUS_TRACK,
 }
 
 // When we start the background service isolate, we only ever enter it once.
@@ -155,6 +161,9 @@ class AudioPlayer {
   final StreamController<String> _errorController =
       StreamController<String>.broadcast();
 
+  final StreamController<PlayerControlCommand> _commandController =
+      StreamController<PlayerControlCommand>.broadcast();
+
   PlayingRouteState _playingRouteState = PlayingRouteState.SPEAKERS;
 
   /// Reference [Map] with all the players created by the application.
@@ -225,6 +234,11 @@ class AudioPlayer {
   ///
   /// Events are sent when an unexpected error is thrown in the native code.
   Stream<String> get onPlayerError => _errorController.stream;
+
+  /// Stream of remote player command send by native side
+  ///
+  /// Events are sent user tap system remote control command.
+  Stream<PlayerControlCommand> get onPlayerCommand => _commandController.stream;
 
   /// Handler of changes on player state.
   @deprecated
@@ -300,7 +314,7 @@ class AudioPlayer {
           PluginUtilities.getCallbackHandle(_backgroundCallbackDispatcher);
       assert(handle != null, 'Unable to lookup callback.');
       _invokeMethod('startHeadlessService', {
-        'handleKey': <dynamic>[handle.toRawHandle()]
+        'handleKey': <dynamic>[handle.toRawHandle()],
       });
     }
   }
@@ -344,7 +358,8 @@ class AudioPlayer {
   /// `callback` is invoked on a background isolate and will not have direct
   /// access to the state held by the main isolate (or any other isolate).
   Future<bool> monitorNotificationStateChanges(
-      void Function(AudioPlayerState value) callback) async {
+    void Function(AudioPlayerState value) callback,
+  ) async {
     if (callback == null) {
       throw ArgumentError.notNull('callback');
     }
@@ -372,6 +387,7 @@ class AudioPlayer {
     bool respectSilence = false,
     bool stayAwake = false,
     bool duckAudio = false,
+    bool recordingActive = false,
   }) async {
     isLocal ??= isLocalUrl(url);
     volume ??= 1.0;
@@ -383,9 +399,51 @@ class AudioPlayer {
       'isLocal': isLocal,
       'volume': volume,
       'position': position?.inMilliseconds,
+      'respectSilence': respectSilence ?? false,
+      'stayAwake': stayAwake ?? false,
+      'duckAudio': duckAudio ?? false,
+      'recordingActive': recordingActive ?? false,
+    });
+
+    if (result == 1) {
+      state = AudioPlayerState.PLAYING;
+    }
+
+    return result;
+  }
+
+  /// Plays audio in the form of a byte array.
+  ///
+  /// This is only supported on Android currently.
+  Future<int> playBytes(
+    Uint8List bytes, {
+    double volume = 1.0,
+    // position must be null by default to be compatible with radio streams
+    Duration position,
+    bool respectSilence = false,
+    bool stayAwake = false,
+    bool duckAudio = false,
+    bool recordingActive = false,
+  }) async {
+    volume ??= 1.0;
+    respectSilence ??= false;
+    stayAwake ??= false;
+
+    if (!Platform.isAndroid) {
+      throw PlatformException(
+        code: 'Not supported',
+        message: 'Only Android is currently supported',
+      );
+    }
+
+    final int result = await _invokeMethod('playBytes', {
+      'bytes': bytes,
+      'volume': volume,
+      'position': position?.inMilliseconds,
       'respectSilence': respectSilence,
       'stayAwake': stayAwake,
       'duckAudio': duckAudio,
+      'recordingActive': recordingActive
     });
 
     if (result == 1) {
@@ -485,24 +543,29 @@ class AudioPlayer {
   /// Sets the notification bar for lock screen and notification area in iOS for now.
   ///
   /// Specify atleast title
-  Future<dynamic> setNotification(
-      {String title,
-      String albumTitle,
-      String artist,
-      String imageUrl,
-      Duration forwardSkipInterval,
-      Duration backwardSkipInterval,
-      Duration duration,
-      Duration elapsedTime}) {
+  Future<dynamic> setNotification({
+    String title,
+    String albumTitle,
+    String artist,
+    String imageUrl,
+    Duration forwardSkipInterval = Duration.zero,
+    Duration backwardSkipInterval = Duration.zero,
+    Duration duration = Duration.zero,
+    Duration elapsedTime = Duration.zero,
+    bool hasPreviousTrack = false,
+    bool hasNextTrack = false,
+  }) {
     return _invokeMethod('setNotification', {
       'title': title ?? '',
       'albumTitle': albumTitle ?? '',
       'artist': artist ?? '',
       'imageUrl': imageUrl ?? '',
-      'forwardSkipInterval': forwardSkipInterval?.inSeconds ?? 30,
-      'backwardSkipInterval': backwardSkipInterval?.inSeconds ?? 30,
-      'duration': duration?.inSeconds ?? 0,
-      'elapsedTime': elapsedTime?.inSeconds ?? 0
+      'forwardSkipInterval': forwardSkipInterval.inSeconds,
+      'backwardSkipInterval': backwardSkipInterval.inSeconds,
+      'duration': duration.inSeconds,
+      'elapsedTime': elapsedTime.inSeconds,
+      'hasPreviousTrack': hasPreviousTrack,
+      'hasNextTrack': hasNextTrack
     });
   }
 
@@ -514,11 +577,16 @@ class AudioPlayer {
   /// this method.
   ///
   /// respectSilence is not implemented on macOS.
-  Future<int> setUrl(String url,
-      {bool isLocal: false, bool respectSilence = false}) {
+  Future<int> setUrl(
+    String url, {
+    bool isLocal: false,
+    bool respectSilence = false,
+  }) {
     isLocal = isLocalUrl(url);
-    return _invokeMethod('setUrl',
-        {'url': url, 'isLocal': isLocal, 'respectSilence': respectSilence});
+    return _invokeMethod(
+      'setUrl',
+      {'url': url, 'isLocal': isLocal, 'respectSilence': respectSilence},
+    );
   }
 
   /// Get audio duration after setting url.
@@ -557,6 +625,7 @@ class AudioPlayer {
       players.remove(playerId);
       return;
     }
+    if (player == null) return;
 
     final value = callArgs['value'];
 
@@ -595,6 +664,12 @@ class AudioPlayer {
         // ignore: deprecated_member_use_from_same_package
         player.errorHandler?.call(value);
         break;
+      case 'audio.onGotNextTrackCommand':
+        player._commandController.add(PlayerControlCommand.NEXT_TRACK);
+        break;
+      case 'audio.onGotPreviousTrackCommand':
+        player._commandController.add(PlayerControlCommand.PREVIOUS_TRACK);
+        break;
       default:
         _log('Unknown method ${call.method} ');
     }
@@ -629,6 +704,7 @@ class AudioPlayer {
     if (!_errorController.isClosed) futures.add(_errorController.close());
 
     await Future.wait(futures);
+    players.remove(playerId);
   }
 
   Future<int> earpieceOrSpeakersToggle() async {