+++ lib/audioplayers_web.dart
@@ -2,7 +2,6 @@ import 'dart:async';
 import 'dart:html';
 
 import 'package:audioplayers/audioplayers.dart';
-
 import 'package:flutter/services.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';